Add literacy/register so requires hook can be optional.
