+++ server.js
@@ -1,5 +1,5 @@
-// Load the Literacy module.
-require('./src');
+// Load the Literacy file extension hook for require.
+require('./src/register');
 
 // And illustrate loading a .js.rst file.
 require('./examples/blocks');
